base_drf/requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
certifi==2020.4.5.1
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
Django==2.2.12
django-allauth==0.42.0
django-appconf==1.0.4
django-ckeditor==5.9.0
django-cors-headers==3.3.0
django-js-asset==1.2.2
django-rest-auth==0.9.5
django-rest-swagger==2.2.0
django-user-accounts==3.0.2
djangorestframework==3.9.2
idna==2.9
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mysqlclient==2.1.0
oauthlib==3.1.0
openapi-codec==1.3.2
Pillow==7.1.2
python3-openid==3.1.0
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.0
simplejson==3.17.0
six==1.15.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.9
